fix: Remove appCheck.ts and fix session extraction issues
- Delete appCheck.ts entirely (no longer needed)
- Update imports in auth.ts and validateTokens.ts to use session.ts directly
- Remove cookie deletion from getSession() to fix Server Component constraint
- Legacy sessions return null and are rejected by proxy middleware

Fixes:
- "Only async functions allowed in 'use server' file" (re-export issue)
- "Cookies can only be modified in Server Action" (getSession in Server Component)
