--- web/src/lib/firebase/actions/appCheck.ts
@@ -1,4 +1,0 @@
-'use server'
-
-// getAppCheckToken is now extracted from JWT session
-export { getAppCheckToken } from '../../auth/session'
