+++ web/src/lib/firebase/actions/validateTokens.ts
@@ -97,8 +97,7 @@ import { getAuth, type Auth } from 'firebase-admin/auth'
 import { getAppCheck } from 'firebase-admin/app-check'
 
 import serviceAccount from '../firebase.json' with { type: 'json' }
-import { getIdToken } from './auth'
-import { getAppCheckToken } from './appCheck'
+import { getIdToken, getAppCheckToken } from '../../auth/session'
 
 /**
  * Result of token validation
